# cmd/api/docs/swagger.yaml
definitions:
...
  entity.Player:
    properties:
      created_at:
        type: string
      deleted_at:
        description: DeletedAt *gorm.DeletedAt `gorm:"index" json:"deleted_at"`
        type: string
      height_cm:
        type: number
      id:
        type: integer
      jersey_number:
        type: integer
      name:
        type: string
      position:
        type: string
      team_id:
        type: integer
      updated_at:
        type: string
      weight_kg:
        type: number
    type: object
  entity.Team:
    properties:
      address:
        type: string
      city:
        type: string
      created_at:
        type: string
      deleted_at:
        description: DeletedAt *gorm.DeletedAt `gorm:"index" json:"deleted_at"`
        type: string
      founded_year:
        type: integer
      id:
        type: integer
      logo_url:
        type: string
      name:
        type: string
      total_wins:
        type: integer
      updated_at:
        type: string
    type: object
  request.AddPlayerRequest:
    properties:
      height_cm:
        type: number
      jersey_number:
        type: integer
      name:
        maxLength: 255
        type: string
      position:
        enum:
        - penyerang
        - gelandang
        - bertahan
        - penjaga_gawang
        type: string
      weight_kg:
        type: number
    required:
    - jersey_number
    - name
    - position
    type: object
  request.AddTeamRequest:
    properties:
      address:
        type: string
      city:
        maxLength: 100
        type: string
      founded_year:
        type: integer
      logo_url:
        type: string
      name:
        maxLength: 255
        type: string
    required:
    - name
    type: object
...
  response.BaseResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
  response.PaginationMetadata:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
  response.PlayerResponse:
    properties:
      message:
        type: string
      player:
        $ref: '#/definitions/entity.Player'
      status:
        type: integer
    type: object
  response.PlayersResponse:
    properties:
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMetadata'
      players:
        items:
          $ref: '#/definitions/entity.Player'
        type: array
      status:
        type: integer
    type: object
  response.TeamResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      team:
        $ref: '#/definitions/entity.Team'
    type: object
  response.TeamsResponse:
    properties:
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMetadata'
      status:
        type: integer
      teams:
        items:
          $ref: '#/definitions/entity.Team'
        type: array
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is the documentation for the main e-commerce service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ayo Test API
  version: "1.0"
paths:
...
  /players/{id}:
    delete:
      description: Delete player by ID (admin only, soft delete)
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Player
      tags:
      - players
    get:
      description: Get player by ID
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PlayerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Player
      tags:
      - players
    put:
      consumes:
      - application/json
      description: Update player by ID (admin only)
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: integer
      - description: Player data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.AddPlayerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PlayerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Player
      tags:
      - players
  /players/team/{teamId}:
    get:
      consumes:
      - application/json
      description: Get all players in a team with pagination
      parameters:
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PlayersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: List Players by Team
      tags:
      - players
    post:
      consumes:
      - application/json
      description: Add a new player to a team (admin only)
      parameters:
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: integer
      - description: Player data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.AddPlayerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PlayerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Player
      tags:
      - players
  /teams:
    get:
      consumes:
      - application/json
      description: Get all teams with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TeamsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: List Teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Add new team (admin only)
      parameters:
      - description: Team data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.AddTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Team
      tags:
      - teams
  /teams/{id}:
    delete:
      description: Delete team by ID (admin only, soft delete)
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Team
      tags:
      - teams
    get:
      description: Get team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TeamResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Team
      tags:
      - teams
    put:
      consumes:
      - application/json
      description: Update team by ID (admin only)
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: Team data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.AddTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Team
      tags:
      - teams
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"